+++ sme_ptrf_apps/core/tasks/__init__.py
@@ -7,4 +7,6 @@ from .gerar_previa_relacao_de_bens import gerar_previa_relacao_de_bens_async
 from .gerar_arquivo_ata import gerar_arquivo_ata_async
 from .gerar_notificacao_inicio_periodo_pc import gerar_notificacao_inicio_periodo_prestacao_de_contas_async
 from .gerar_notificacao_pendencia_envio_pc import gerar_notificacao_pendencia_envio_prestacao_de_contas_async
+from .gerar_notificacao_proximidade_inicio_pc import gerar_notificacao_proximidade_inicio_periodo_prestacao_conta_async
+
